Added nU Version 240322 and µU Version 240104.

+++ nU/dgra/changRoberts.go
@@ -1,11 +1,11 @@
 package dgra
 
-// (c) Christian Maurer   v. 231220 - license see nU.go
+// (c) Christian Maurer   v. 231228 - license see nU.go
 
 func (x *distributedGraph) ChangRoberts() {
   x.connect(uint(0))
   defer x.fin()
-  out, in := uint(0), uint(1)
+  out, in := 0, 1
   if x.Graph.Outgoing(1) { in, out = out, in }
   x.ch[out].Send (x.me)
   for {
@@ -18,7 +18,7 @@ func (x *distributedGraph) ChangRoberts() {
         x.ch[out].Send (inf + x.me)
         return
       }
-    } else { // n > inf
+    } else { // id >= inf
       x.leader = id - inf
       if x.leader != x.me {
         x.ch[out].Send (id)
